requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.11.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
networkx==3.4.2
numpy==2.2.6
openai==1.82.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pypdf==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.0
zstandard==0.23.0
